search engine optimization
Web Crawler search engine is a meta search engine which is to be
developed using             Apache Lucene search engine library. A Meta
search engine transmits user’s search simultaneously to several
individual search engines and their databases of web pages and gets
results from all the search engines queried. We could thus save a lot
of time by initiating the search at a single point and sparing the need
to use and learn several separate search engines. This can be even more
helpful, if we are looking for a broad range of results.